Nsw edm fixes

See merge request atlas/athena!14600

(cherry picked from commit d799e88ebf4b88a944ab454ec4c98b2484ed8614 [formerly 939967bd5211420809516169adf9684914843fae])

bc772770 Fix in STGC TP converters
707a8999 Fixes in the NSW EDM
f04a5a85 Fixes in nsw EDM
77d95bd7 Fixes in sTGC PRD and ntuple dump algorithm
4f96f8f5 Moving sTGC RDO charge to uint16_t from float

Former-commit-id: 5d48944195edd1653d45437e7c6c64f7a1b36b8c

+++ MuonSpectrometer/MuonValidation/MuonPRDTest/src/sTGCRDOVariables.cxx
@@ -118,7 +118,7 @@ StatusCode sTGCRDOVariables::initializeVariables()
   m_NSWsTGC_rdo_channel     = new std::vector<int>();
   m_NSWsTGC_rdo_channel_type= new std::vector<int>();
   m_NSWsTGC_rdo_time        = new std::vector<double>();
-  m_NSWsTGC_rdo_charge      = new std::vector<double>();
+  m_NSWsTGC_rdo_charge      = new std::vector<uint16_t>();
   m_NSWsTGC_rdo_bcTag       = new std::vector<uint16_t>();
   m_NSWsTGC_rdo_isDead      = new std::vector<bool>();
